관리자페이지 비즈니스로직 추가

+++ src/main/java/com/project/service/ProductService.java
@@ -1,18 +1,25 @@
 package com.project.service;
 
 import java.util.List;
+import java.util.Objects;
 
 import com.project.entity.ProductEntity;
+import com.project.error.NotFoundException;
 import com.project.repository.ProductRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class ProductService {
 
     @Autowired
     ProductRepository repository;
+
+    public List<ProductEntity> findAll() {
+        return repository.findAll();
+    }
 
     /**
      * 페이지와 한 페이지당 표시할 단위의 상품 목록을 가져옴
@@ -34,6 +41,12 @@ public class ProductService {
         return repository.countAllProduct();
     }
 
+    /**
+     * 상품 추가
+     * 
+     * @param product
+     * @return
+     */
     public int addProduct(ProductEntity product) {
         try {
             repository.save(product);
@@ -44,14 +57,19 @@ public class ProductService {
         return 1;
     }
 
+    /**
+     * 상품 삭제
+     * 
+     * @param productNo
+     * @return
+     */
+    @Transactional
     public int deleteProduct(Integer productNo) {
-        try {
-            ProductEntity product = repository.findByProductNo();
-            repository.delete(product);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return 0;
+        ProductEntity product = repository.findByProductNo(productNo);
+        if (Objects.isNull(product)) {
+            throw new NotFoundException("not found user");
         }
+        repository.delete(product);
         return 1;
     }
 }
